Recommend turning Yellow off before connecting via serial
I spent a long time trying to figure out why no new USB devices were showing up on my MacBook. It turns out it's because I was waiting for the Yellow to start up before connecting it via USB. Apparently it must be connected when the Yellow _starts_ in order for it to work.